添加stream output
