+++ package.json
@@ -25,8 +25,5 @@
 		"obsidian": "^1.4.0",
 		"tslib": "2.3.1",
 		"typescript": "^4.9.5"
-	},
-	"dependencies": {
-		"@codemirror/state": "^6.5.0"
 	}
 }
